requirements.txt
aiodns==3.6.0
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
Django==4.2.26
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
Faker==38.2.0
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
